edits

+++ fileui.c
@@ -1,3 +1,4 @@
+#include "fileui.h"
 #include <stdio.h>      //printf
 #include <sys/stat.h>   //mkdir
 #include <sys/types.h>  //mode_t (0777)
@@ -10,44 +11,17 @@ void file_simulation()
     printf("Type 'help' for a list of commands, or 'exit' to quit.\n");
 }
 
-/* example of what can be printed 
+/* example of what can be printed, add these 
 Welcome to File System Simulation
 -----------------------------------
 $> help
 Available commands:
-- ls       - List directory contents
-- mkdir    - Create a new directory
 - touch    - Create a new file
 - rm       - Remove a file
-- cd       - Change directory
 - cat      - Display file contents
-- exit     - Exit the simulation
-
-$> ls
-Directory: /
-- home/
-- user1.txt
-- docs/
-
-$> mkdir projects
-Created directory: projects
-
-$> ls
-Directory: /
-- home/
-- user1.txt
-- docs/
-- projects/
-
-$> cd projects
-Changed directory to: projects/
 
 $> touch readme.txt
 Created file: readme.txt
-
-$> ls
-Directory: projects/
-- readme.txt
 
 $> cat readme.txt
 File is empty.
@@ -55,25 +29,26 @@ File is empty.
 $> echo "This is a project!" > readme.txt
 Added content to readme.txt
 
-$> cat readme.txt
-This is a project!
-
 $> rm readme.txt
 Removed file: readme.txt
 
-$> ls
-Directory: projects/
-- (empty)
-
-$> cd ..
-Changed directory to: /
-
-$> exit
-Exiting the simulation. Goodbye!
 */
 void list_contents()
 {
+    DIR *dir;
+    struct dirent *entry;
 
+    dir = opendir(".");
+    if (dir == NULL) {
+        perror("Unable to open current directory");
+        return;
+    }
+    while ((entry = readdir(dir)) != NULL) {
+        if (strcmp(entry->d_name, ".") != 0 && strcmp(entry->d_name, "..") != 0) {
+            printf("- %s\n", entry->d_name);
+        }
+    }
+    closedir(dir);
 }
 
 void create_directory(const char *name)
@@ -92,12 +67,49 @@ void create_directory(const char *name)
 void delete_directory(const char *name)
 {
     const char *directory_name = name;
-    if(rmdir(directory_name) == 0)
+    //check if directory exists 
+    DIR *dir = opendir(directory_name);
+    if(dir)
     {
-        printf("Directory succesfully deleted.\n");
+        closedir(dir);
+        if(rmdir(directory_name) == 0)
+        {
+            printf("Directory succesfully deleted.\n");
+        }
+        else
+        {
+            perror("rmdir failed");
+        }
     }
     else
     {
-        perror("rmdir failed");
+        perror("Directory does not exist");
     }
+}
+
+void change_directory(const char *name)
+{
+    const char *directory_name = name;
+    DIR *dir = opendir(directory_name);
+    if(dir)
+    {
+        closedir(dir);
+        if(chdir(directory_name)==0)
+        {
+            char cwd[1024]; //stores full path to current directory 
+            if(getcwd(cwd, sizeof(cwd)) != NULL)
+            {
+                printf("Directory changed.\n");
+            }
+        }
+        else
+        {
+            perror("cd failed");
+        }
+    }
+    else
+    {
+        perror("Directory does not exist.\n");
+    }
+    
 }
